+++ ripe/atlas/tools/commands/report.py
@@ -123,11 +123,7 @@ class Command(BaseCommand):
         if self.arguments.aggregate_by:
             results = aggregate(results, self.get_aggregators())
 
-        Rendering(
-            renderer=renderer,
-            header=self._get_header(measurement),
-            payload=results
-        ).render()
+        Rendering(renderer=renderer, payload=results).render()
 
     def get_aggregators(self):
         """Return aggregators list based on user input"""
@@ -145,17 +141,3 @@ class Command(BaseCommand):
             else:
                 aggregation_keys.append(aggregation_class(key=key))
         return aggregation_keys
-
-    def _get_header(self, measurement):
-        """
-        Most of the time you want a fancy header, but for the raw renderer,
-        we want nothing.
-        """
-
-        description = measurement.description or ""
-        if description:
-            description = "\n{}".format(description)
-
-        return ("\nRIPE Atlas Report for Measurement #{}\n"
-                "==================================================="
-                "{}\n".format(measurement.id, description))